+++ R/utils_threads.R
@@ -1,18 +1,19 @@
 LIDRTHREADS = new.env()
 LIDRTHREADS$n = 1L
+LIDRTHREADS$input <- 1L
 
 #' Set or get number of threads that lidR should use
 #'
 #' Set and get number of threads to be used in lidR functions that are parallelized with OpenMP.
-#' Default value 0 means to utilize all CPU available. \code{get_lidr_threads()} returns the number
+#' 0 means to utilize all CPU available. \code{get_lidr_threads()} returns the number
 #' of threads that will be used. This affects \code{lidR} package but also the \code{data.table} package
 #' by internally calling \link[data.table:openmp-utils]{setDTthreads} because several functions  of
 #' lidR rely on \code{data.table} but it does not change R itself or other packages using OpenMP.
 #'
 #' @seealso \link{lidR-parallelism}
 #'
-#' @param threads An integer >= 0. Default 0 means use all CPU available and leave the operating system
-#' to multi task.
+#' @param threads Positve scala. Default 0 means use all CPU available. Values > 1 mean
+#' using n cores, values in ]0, 1[ mean using a fraction of the cores e.g. 0.5 = half.
 #' @export
 set_lidr_threads = function(threads)
 {
@@ -24,6 +25,7 @@ set_lidr_threads = function(threads)
   if (max < 0)
   {
     if (threads > 1) message("This installation of lidR has not been compiled with OpenMP support") # nocov
+    LIDRTHREADS$input <- 1L
     LIDRTHREADS$n <- 1L # nocov
     data.table::setDTthreads(1L) # nocov
     return(invisible()) # nocov
@@ -31,11 +33,19 @@ set_lidr_threads = function(threads)
 
   if (threads == 0 | threads > max)
   {
+    LIDRTHREADS$input <- threads
     LIDRTHREADS$n <- max
     data.table::setDTthreads(0L)
   }
+  else if(threads < 1)
+  {
+    LIDRTHREADS$input <- threads
+    LIDRTHREADS$n <- as.integer(max*threads)
+    data.table::setDTthreads(as.integer(max*threads))
+  }
   else
   {
+    LIDRTHREADS$input <- as.integer(threads)
     LIDRTHREADS$n <- as.integer(threads)
     data.table::setDTthreads(threads)
   }