+++ src/gcc_4_9_patch/std_is_trivially_substitutes.hpp
@@ -1,0 +1,23 @@
+#ifndef ISTRIVIALLYPATCH
+#define ISTRIVIALLYPATCH
+
+namespace std
+{
+  template <typename T>
+  class is_trivially_copyable
+  {
+  public:
+    bool operator()() { return value; };
+    static constexpr bool value = true;
+  };
+
+  template <typename T>
+  class is_trivially_constructible
+  {
+  public:
+    bool operator()() { return value; };
+    static constexpr bool value = true;
+  };
+}
+
+#endif