+++ NEWS.md
@@ -8,6 +8,7 @@ The parallel evaluation strategy was no recognized and lidR does not know if Ope
 OpenMP has been disabled by security. 
 Use options(lidR.check.nested.parallelism = FALSE) and set_lidr_threads() for a fine control of parallelism.
 ```
+- Fix: incorrect offset computation in `spTransform()` have for consequences to make the function failing with error: `Non quantizable value outside the range of representable values of type 'int'`.
 - Doc: documentation of `options(lidR.check.nested.parallelism = FALSE)` was missing. Information can now be found in `?lidR-package` and `?lidR-parallelism`
 - Enhance: in `catalog_apply()` if `lidR.check.nested.parallelism = FALSE` it now respects the input of `set_lidr_thread()` instead of the output of `get_lidr_threads()`. For example if `set_lidr_thread(0)` it now propagates the information 0 (all cores) instead of the output of `get_lidr_thread()` which might be e.g. 4 on the master worker but might be different on the slave workers. Similarly `set_lidr_thread(20)` will request 20 cores to the workers even if `get_lidr_thread()` returns 4 on the local machine.
 - Enhance: `set_lidr_thread()` accepts inputs < 1 such as 0.5 to mean 'half' or 'quarter' of available cores.